Написал в ядре тестовый код. Сделал запись/чтение бандла в/из файловой системы

+++ src/main/java/dataAccess/entity/Group.java
@@ -74,7 +74,7 @@ public class Group implements Serializable
 		return id;
 	}
 
-	public void setId(int id)
+	public void setId(long id)
 	{
 		this.id = id;
 	}
